Fix for async code running in background
This fix applies an "await" to two calls to the "start" function within
the "reverse_game.js" file which were causing a double output while
playing the game. Since these functions were run in the background and
were not awaited, their output doubled up to the console at the same
time, making it appear that the start function logged twice, when it was
a function being called twice without being awaited.